pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "edgar-parser"
version = "0.3.0"
description = "Structured financial data from SEC EDGAR filings — 10-Q, 10-K, and 8-K"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "pandas",
    "numpy",
    "beautifulsoup4",
    "lxml",
    "openpyxl",
    "python-dateutil",
    "rapidfuzz",
    "python-dotenv",
]

[project.urls]
Homepage = "https://github.com/henrysouchien/edgar-parser"
Source = "https://github.com/henrysouchien/edgar-parser"
Issues = "https://github.com/henrysouchien/edgar-parser/issues"

[project.optional-dependencies]
llm = ["anthropic", "openai"]

[tool.hatch.build.targets.wheel]
packages = ["edgar_parser"]

[tool.hatch.build.targets.sdist]
include = ["edgar_parser/", "pyproject.toml", "README.md", "LICENSE"]
